W460: structural-gap census; +3 MATCH (6,925 -> 6,928), nd -2,799
New instrument: tools/gap_census.py aligns the candidate and retail instruction
streams with difflib (relocated fields neutralised first) and reports every
block only one side emits.  A retail-only block is code retail has and we do
not; a candidate-only block is code we emit and retail does not.  Neither can
be closed by register or declaration tuning, and each displaces every later
comparison, which is why the functions carrying them are the ones that never
move.  Blocks are tagged LOOPSHAPE / CALL / SPILL / ZEROINIT / FPU / BODY.

Catalogue-wide: 1,133 functions, 176,143 retail-only words against 165,158
candidate-only words.  **85 functions are within 8 words of structurally
complete** — the highest-probability MATCH pool the campaign has had.

Gate (build/baselines/W460_gate.json vs W459_gate.json):

  first-party MATCH      6,925 -> 6,928   (+3)
  first-party nd         1,195,112 -> 1,192,313   (-2,799)
  MATCH lost             0
  functions improved     30
  functions regressed    0
  over-window            3 -> 3
  nd-0 but not MATCH     0
  pytest                 92 passed
  build                  SLUS_216.21 sha1 3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb  OK
  reloc_audit            0 wrong call targets

New MATCHes:

  btlEffect func_002bb3a0   202 -> 0
  btlEffect func_002c65d0   106 -> 0
      A typed u32 alias for datCalcRand removed a spurious dsll/dsra pair, a
      bool local became u32, and a static inline forced retail's addu operand
      order.
  y_misc FUN_00418c70        29 -> 0
      Whole-object `matrix.pos = *(RwV3d *)afStack_b0;` fixed the access
      ordering; Add/Mul inline helpers fixed FPU operand order; declaration
      order fixed the copy loop's register colouring.

Largest other gains:

  op_menu  opMenu0026d430   658 -> 124   item-status address materialised as u32*
  bpc_main FUN_002441b0     698 -> 350   if(x==1) rewritten as a single-case switch
  k_clump  func_001a8140   1045 -> 714   explicit _abs render state + absolute camera
  bai_main FUN_0024c110    1492 -> 1222  direct (u8)(255.0f * fade / baseY)
  br_root  func_001f0c40    161 ->  21
  datCalc  FUN_00306610     338 ->  202  recovered six-argument FUN_00306020 ABI
  datCalendar func_00183840 520 ->  406
  fclShopMisc FUN_003f9220  383 ->  262

Cross-TU ABI recovered: FUN_00388000 is
`(float z, float alpha, float *pos, float *size, float *uv0, float *uv1,
  u32 *texturePtr, RwRGBA *color, u32 flags)`, derived from retail's
$a0-$a3/$t0/$t1/$t2/$f12/$f13 mapping.  The missing second float had shifted
every pointer one slot.  Corrected across mt_evtMisc, mt_evtMain, mt_evtMenu
and mt_evtMessage; the _2f alias and the dead _d1c0 alias are gone.

Two mechanisms measured to their boundary — both negative, both recorded so the
next wave does not repeat them:

1. **LOOPSHAPE is not a licence to change the loop keyword.**  W459 got two
   MATCHes converting do/while to a pre-test while.  This wave two lanes tested
   four more: fclCombineList FUN_003dffc0 went 15 -> 107, bcm_main FUN_001FFC60
   347 (from 295), and both bp_tex FUN_0022B630 loops were flat.  A retail-only
   branch at a loop head can equally be the same loop with a different entry
   sequence.  The census tag locates the defect; it does not name the fix.

2. **h_camp FUN_001368a0 (nd 8) is a genuine floor.**  Sixteen source-level
   probes: every parameter permutation, both declaration orders, union member
   order, a helper, and both sink directions.  All flat at 8 or worse.  Retail's
   prologue parameter save order is not reachable from source there.

Also corrected: sfl_ground func_002392d0 has been labelled "99.2% implemented,
call census CLEAN" for several waves.  It is not.  Retail's grid body makes four
jals where ours makes three, and retail's $f30 (an angle accumulated across
three sin calls) and $f31 (a distance ratio read five times, including a madd.s)
have no equivalent in our source at all.  The two missing callee-saved FPU
registers are a consequence of missing per-tile code.  Reclassified as a
reconstruction target.

Rejected at ruling time, for correctness rather than measurement:

  - k_vpad func_001e8e80: "merging duplicate mode==1/mode==2 branches" took nd
    2019 -> 1453 while the object went 2,688 -> 2,004 against a 2,720 window.
    Retail emits the duplication; 648 bytes of real code had been deleted.
    nd falls when an object shrinks because there is less left to differ.
  - sfl_cursor FUN_0025a130: a `volatile` aggregate for 6 nd.  Hard ban.
  - bsa_main: a `#define D_0096009C_abs (*renderQuad)` shadowing a global for
    the rest of the TU, plus three deleted statements, left behind by an
    incomplete revert.  Restored from the commit.
  - btlAction btlActionUpdateStateError: a dropped `btlAction0028a780(action);`
    found by diffing the whole changed-file set at gate time (nd 308 -> 497).
  - h_camp FUN_001432f0: converting five of nine identical constant pairs
    measured 1,249, but the five were chosen by which ones improved.  All nine
    measures 1,349, worse than the 1,315 baseline.  Reverted; a construct
    selected by the metric is fitted, not reconstructed.

Lane reports: build/wip/*W460_report.md.

+++ src/Main/Battle/Cmd/bai_main.c
@@ -127,7 +127,7 @@ void FUN_0024c110(void)
     color[0] = 0xff;
     color[1] = 0xff;
     color[2] = 0xff;
-    color[3] = (u8)(u32)(255.0f * fade);
+    color[3] = (u8)(255.0f * fade);
     func_0021d950(work + 0x110, color);
 
     if (*(s32*)(work + 0xc) < 10) {
@@ -195,7 +195,7 @@ void FUN_0024c110(void)
     color[0] = 0xff;
     color[1] = 0xff;
     color[2] = 0xff;
-    color[3] = (u8)(u32)(255.0f * baseY);
+    color[3] = (u8)(255.0f * baseY);
     func_0021d950(work + 0x210, color);
     timer = *(s32*)(work + 0xc);
 
@@ -258,7 +258,7 @@ void FUN_0024c110(void)
     *(f32*)(work + 0x374) = (f32)color[1];
     *(f32*)(work + 0x378) = (f32)color[2];
     *(f32*)(work + 0x37c) = (f32)color[3];
-    color[3] = (u8)(u32)(255.0f * baseX);
+    color[3] = (u8)(255.0f * baseX);
     *(f32*)(work + 0x3b0) = (f32)color[0];
     *(f32*)(work + 0x3b4) = (f32)color[1];
     *(f32*)(work + 0x3b8) = (f32)color[2];
